Implement SolidusAfterpay::CallbacksController
This controller is responsible for handling the #confirm or #cancel callbacks
from Afterpay.

Consumer completes the Afterpay screenflow and is returned to the Merchant
website:
- if the Consumer clicks "confirm", they will be returned to the Merchant
  website with the orderToken and a status of "SUCCESS".
- if the Consumer cancels, they will be returned to the Merchant website with
  the orderToken and a status of "CANCELLED".

The #confirm action updates the order with the Afterpay payment method and
creates a new payment source storing the orderToken in it. Then it advances the
order to the next state.