Introduce the use_solidus_api preference
Until now the `SolidusAfterpay::Api::CheckoutsController` only handles a
cookie-based authentication.
If a store wants to use an API-based authentication, now it can set the
`use_solidus_api` preference to `true`.
Under the hood that preference changes the parent class of the controller from
`SolidusAfterpay::BaseController` to `::Spree::Api::BaseController`.