Build the shipping rate array for Afterpay
Add the ability to customize the
default shipping rate builder service class.
This class is needed to build an Afterpay shipping rate array based on
the order address.

The Afterpay integrated express checkout delegates to
Afterpay to change the shipping address.
When the shipping address changes, Solidus should recalculate
the shipping rates and send it to Afterpay again.

Co-authored-by: garciajordy <jordygarcia0119@gmail.com>