Fix minimum and maximum amount issue
Add cache for the minimum, maximum and currency values with an expiry
value of one day as Afterpay recommend that, otherwise it will make
an API call for every user trying to checkout.

Add `attr_writer` for the `cache_expiry` to make it configurable with
a default value of `1.day`.