Update the express checkout complete respose
Allow to respond with the solidus order on the express checkout
complete action if the solidus_api are enabled.

This is needed since the express callbacks controller create action
updates the Solidus order information and those information
could be useful to update the custom frontend without making
another API call to Solidus.